🚚 rename : CandidateCancelModal 컴포넌트 domain 폴더구조로 변경
domain 폴더 구조로 변경 완료
