+++ src/components/domain/meet/detail/index.ts
@@ -1,1 +1,2 @@
 export { default as CandidateBottomSheet } from '../detail/CandidateBottomSheet'
+export { default as CandidateCancelModal } from '../detail/CandidateCancelModal'
